+++ CHANGELOG.rst
@@ -7,6 +7,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+[v0.11.1] - 2023-11-22
+---------------------
+
+Fixed
+~~~~~
+
+* reference system parsing for `gmx:Anchor` elements
+
 
 [v0.11.1] - 2023-11-22
 ---------------------